Cleanup, refactor deployment code
